+++ packages/nx-builders/src/builders/run/main.interface.ts
@@ -1,4 +1,4 @@
-import type { ExecaArguments } from '@webundsoehne/nx-tools'
+import type { EnvironmentVariables, ExecaArguments } from '@webundsoehne/nx-tools'
 
 /**
  * Options for run builder
@@ -23,6 +23,9 @@ export interface RunBuilderOptions {
   /** run with node */
   node?: boolean
 
+  /** strictly execute this with node eventhough it can be a node binary as well */
+  executeWithNode?: boolean
+
   /** pass in node options when running as node */
   nodeOptions?: string
 
@@ -30,7 +33,7 @@ export interface RunBuilderOptions {
   watch?: boolean
 
   /** environment variables */
-  environment?: Record<string, string | number>
+  environment?: EnvironmentVariables
 }
 
 export type NormalizedRunBuilderOptions = ExecaArguments