+++ packages/nx-tools/src/utils/file-system/node-bin.interface.ts
@@ -1,0 +1,12 @@
+import type { EnvironmentVariables } from '@interfaces/environment-variables.interface'
+
+export interface NodeBinaryPathExtensionsOptions {
+  env?: EnvironmentVariables
+  start?: string
+  top?: string
+}
+
+export interface NodeBinaryPathExtensions {
+  key: string
+  path: string
+}