+++ packages/nx-nest/CHANGELOG.md
@@ -1,3 +1,54 @@
+# @webundsoehne-private/nx-nest [4.2.0-beta.6](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne-private/nx-nest@4.2.0-beta.5...@webundsoehne-private/nx-nest@4.2.0-beta.6) (2022-02-17)
+
+### Dependencies
+
+- **@webundsoehne/nx-builders:** upgraded to 5.2.0-beta.4
+- **@webundsoehne/nx-tools:** upgraded to 5.2.0-beta.3
+
+# @webundsoehne-private/nx-nest [4.2.0-beta.5](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne-private/nx-nest@4.2.0-beta.4...@webundsoehne-private/nx-nest@4.2.0-beta.5) (2022-02-16)
+
+### Bug Fixes
+
+- add force node flag for builder for oebb ([23a97ce](https://gitlab.tailored-apps.com/ckilic/nx-test/commit/23a97ce8b1de40d3dd98f05ffbd8ed88282d7327))
+
+### Dependencies
+
+- **@webundsoehne/nx-builders:** upgraded to 5.2.0-beta.3
+- **@webundsoehne/nx-tools:** upgraded to 5.2.0-beta.2
+
+# @webundsoehne-private/nx-nest [4.2.0-beta.4](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne-private/nx-nest@4.2.0-beta.3...@webundsoehne-private/nx-nest@4.2.0-beta.4) (2022-02-16)
+
+### Bug Fixes
+
+- swap typeorm-seeding for the supported version ([2a2edaa](https://gitlab.tailored-apps.com/ckilic/nx-test/commit/2a2edaa3e649afbf1c51788bcdf2f24857ed41e5))
+
+# @webundsoehne-private/nx-nest [4.2.0-beta.3](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne-private/nx-nest@4.2.0-beta.2...@webundsoehne-private/nx-nest@4.2.0-beta.3) (2022-02-16)
+
+### Dependencies
+
+- **@webundsoehne/nx-builders:** upgraded to 5.2.0-beta.2
+
+# @webundsoehne-private/nx-nest [4.2.0-beta.2](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne-private/nx-nest@4.2.0-beta.1...@webundsoehne-private/nx-nest@4.2.0-beta.2) (2022-02-16)
+
+### Bug Fixes
+
+- start script ([0c83063](https://gitlab.tailored-apps.com/ckilic/nx-test/commit/0c83063b41d5b74743f3ce293bf64fe2447eac85))
+
+# @webundsoehne-private/nx-nest [4.2.0-beta.1](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne-private/nx-nest@4.1.0...@webundsoehne-private/nx-nest@4.2.0-beta.1) (2022-02-15)
+
+### Bug Fixes
+
+- update environment variables to always be string ([110cb57](https://gitlab.tailored-apps.com/ckilic/nx-test/commit/110cb574aeab31576c347ebe55cbaea73282664c))
+
+### Features
+
+- use npx or yarn exec like paths instead of hardcoding for windows ([d1172a7](https://gitlab.tailored-apps.com/ckilic/nx-test/commit/d1172a7ebe345347ba72b8ffa24f544cfeef56c1))
+
+### Dependencies
+
+- **@webundsoehne/nx-builders:** upgraded to 5.2.0-beta.1
+- **@webundsoehne/nx-tools:** upgraded to 5.2.0-beta.1
+
 # @webundsoehne-private/nx-nest [4.1.0](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne-private/nx-nest@4.0.8...@webundsoehne-private/nx-nest@4.1.0) (2022-02-05)
 
 ### Features