+++ packages/patch-package/CHANGELOG.md
@@ -1,3 +1,9 @@
+# @webundsoehne/patch-package [3.2.0-beta.1](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne/patch-package@3.1.0...@webundsoehne/patch-package@3.2.0-beta.1) (2022-02-15)
+
+### Features
+
+- use npx or yarn exec like paths instead of hardcoding for windows ([d1172a7](https://gitlab.tailored-apps.com/ckilic/nx-test/commit/d1172a7ebe345347ba72b8ffa24f544cfeef56c1))
+
 # @webundsoehne/patch-package [3.1.0](https://gitlab.tailored-apps.com/ckilic/nx-test/compare/@webundsoehne/patch-package@3.0.2...@webundsoehne/patch-package@3.1.0) (2022-02-05)
 
 ### Features